requirements.txt
flask
google-generativeai
python-dotenv
Flask-CORS